Exposed eventLoop in the Nova class.
New duplicate event handling: IDProviders can be registered at the EventLoop which allow a new event to override an existing one in the RingBuffer, *BEFORE* it gets put into it. We do no longer rely on the BatchProcessing for that. This allows us to protect against ringBuffer overruns by throwing away outdated stuff.